requirements.txt
Django==1.4.2
argparse==1.2.1
distribute==0.6.28
dj-database-url==0.2.1
psycopg2==2.4.5
wsgiref==0.1.2
django-registration==0.8
